根据annotation注解修改bean注入
