--- mavenPro/src/main/resources/application.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-        https://www.springframework.org/schema/beans/spring-beans.xsd">
-
-    <bean id="userService" class="com.liu.learnjava.service.UserService">
-        <property name="mailService" ref="mailService" />
-    </bean>
-    <bean id="mailService" class="com.liu.learnjava.service.MailService" />
-<!--    注入是Boolean，int，string这种数据类型的写法-->
-<!--    <bean id="dataSource" class="com.zaxxer.hikari.HikariDataSource">-->
-<!--        <property name="jdbcUrl" value="jdbc:mysql://localhost:3306/test" />-->
-<!--        <property name="username" value="root" />-->
-<!--        <property name="password" value="password" />-->
-<!--        <property name="maximumPoolSize" value="10" />-->
-<!--        <property name="autoCommit" value="true" />-->
-<!--    </bean>-->
-</beans>
